working spotify query, need to connect to backend
