+++ .gitignore
@@ -1,3 +1,3 @@
 node_modules
-bundle.js
-
+public/bundle.js
+keys.js
